Rename files to match the new content
We don't plan to add more predicates for now, so we rename
`_predicates.py` to `_with_previous.py` and
`test_predicates_only_if_previous.py` to `test_with_previous.py`.
